+++ src/constants/workflows.js
@@ -1,0 +1,6 @@
+const WORKFLOW_TYPES = [
+  {value: 'request', text: 'request'},
+  {value: 'opportunity', text: 'opportunity'}
+];
+
+export default WORKFLOW_TYPES